Add source_code_hash variable
When using this module in combination with `data.archive_file` it fails
because `filebase64sha256()` is called before the data resource is
refreshed.

By using the `output_base64sha256` output from `data.archive_file` we
get around this, but need a way to pass in this value, thus introduce
the `source_code_hash` variable to optionally set the value and if not
for the module to compute it.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>